+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Minecraft 26.1.1 and 26.1.2 support (Fabric, NeoForge, Forge)
+
 ## [0.2.0] - 2026-04-05
 
 ### Added